Moved AddTransaction and Quote logic to seperate components - Used typescript to create the component - Added typescript type for functions it makes the code more easily - understandable and readable
