+++ src/components/Quote.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Quote: React.FC = () => {
+    return (
+        <div className="quote">You should know where you are spending</div>
+    )
+}
+
+export default Quote;
